+++ demo-sse-jaxrs-client/src/main/java/com/example/demo/dao/SseLastEventIdRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.dao;
 import com.example.demo.dao.entity.SseLastEventId;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SseLastEventIdRepository extends CrudRepository<SseLastEventId, Integer> {
+public interface SseLastEventIdRepository extends CrudRepository<SseLastEventId, Long> {
     SseLastEventId findByName(String name);
 }
